# services/booking-service/src/main/java/ru/larkin/bookingservice/dto/req/UpdateUserRequest.java
package ru.larkin.bookingservice.dto.req;

import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.Size;

public class UpdateUserRequest {

    @Size(min = 1)
    private String username;

    @Email
    private String email;

    @Size(min = 8)
    private String password;

    private RoleEnum role;

    public enum RoleEnum {
        ADMIN,
        USER
    }

    public String getUsername() {
        return username;
    }

    public void setUsername(String username) {
        this.username = username;
    }

    public String getEmail() {
        return email;
    }

    public void setEmail(String email) {
        this.email = email;
    }

    public String getPassword() {
        return password;
    }

    public void setPassword(String password) {
        this.password = password;
    }

    public RoleEnum getRole() {
        return role;
    }

    public void setRole(RoleEnum role) {
        this.role = role;
    }
}