# services/booking-service/src/main/java/ru/larkin/bookingservice/dto/resp/SuccessResponse.java
package ru.larkin.bookingservice.dto.resp;

import jakarta.validation.Valid;
import jakarta.validation.constraints.NotNull;
import java.time.OffsetDateTime;

public class SuccessResponse {

    private Integer status;

    private String message;

    @NotNull
    @Valid
    private OffsetDateTime timestamp;

    public Integer getStatus() {
        return status;
    }

    public void setStatus(Integer status) {
        this.status = status;
    }

    public String getMessage() {
        return message;
    }

    public void setMessage(String message) {
        this.message = message;
    }

    public OffsetDateTime getTimestamp() {
        return timestamp;
    }

    public void setTimestamp(OffsetDateTime timestamp) {
        this.timestamp = timestamp;
    }
}